+++ 02-data-modeling-cassandra/README.md
@@ -1,0 +1,5 @@
+# Data Modeling with Apache Cassandra
+
+In this project, we examine NoSQL database technology to build the star schema.
+
+
