+++ 03-data-modeling-aws-data-warehouse/sample-dwh.cfg
@@ -1,0 +1,23 @@
+[AWS]
+KEY=
+SECRET=
+
+[CLUSTER]
+REGION=<cluster_region>
+TYPE=multi-node
+NAME=<cluster_name>
+NUM_NODES=<cluster_number_of_nodes>
+NODE_TYPE=dc2.large
+
+DB=<database_name>
+DB_USER=<database_user>
+DB_PASSWORD=<password>
+DB_PORT=<database_port>
+
+[IAM_ROLE]
+ARN=<iam_role_arn_name>
+
+[S3]
+LOG_DATA='s3://udacity-dend/log_data'
+LOG_JSONPATH='s3://udacity-dend/log_json_path.json'
+SONG_DATA='s3://udacity-dend/song_data'
