+++ 04-data-lakes-with-spark/dl.cfg
@@ -1,0 +1,3 @@
+[default]
+AWS_ACCESS_KEY_ID=
+AWS_SECRET_ACCESS_KEY=
