+++ 04-data-lakes-with-spark/requirements.txt
@@ -1,0 +1,1 @@
+pyspark>=3.0.2
