+++ README.md
@@ -1,3 +1,55 @@
-Data Engineering Nano Degree Program from Udacity
+# Sparkify Data Engineering Projects for Song Play Analytics
+
+This repo explores different big data technologies for analyzing songs and user
+activity data captured by the music streaming app of the ficticious Sparkify startup.
+
+The goal is to examine a number of technologies to a build business intelligence
+platform from which the data scientists can easily analyze the songs and user activity
+data collected on the app.
+
+## Datasets
+
+The Sparkify music streaming app collects song metadata and user activity on
+the app in json log files.  The sample of these log files are stored under */data*
+directory.  
+
+### Song Metadata Dataset
+The song dataset is a subset of real data from the [Million Song Dataset](http://millionsongdataset.com/). 
+Each file is in JSON format and contains metadata about a song and the artist 
+of that song. The files are partitioned by the first three letters of each song's 
+track ID. For example, here are filepaths to two files in this dataset.
+
+```
+data/song_data/A/B/C/TRABCEI128F424C983.json
+data/song_data/A/A/B/TRAABJL12903CDCF1A.json
+```
+And below is an example of what a single song file, TRAABJL12903CDCF1A.json, looks like.
+
+```
+{"num_songs": 1, "artist_id": "ARJIE2Y1187B994AB7", "artist_latitude": null, "artist_longitude": null, "artist_location": "", "artist_name": "Line Renaud", "song_id": "SOUPIRU12A6D4FA1E1", "title": "Der Kleine Dompfaff", "duration": 152.92036, "year": 0}
+```
+
+### User Activity Log Dataset
+The second database consists of log files in JSON format generated by this 
+[event simulator](https://github.com/Interana/eventsim) based on the songs 
+in the dataset above. These simulate activity logs from a music streaming app 
+based on specified configurations.
+
+The log files in the dataset are partitioned by year and month. For example, 
+here are filepaths to two files in this dataset.
+
+```
+data/log_data/2018/11/2018-11-12-events.json
+data/log_data/2018/11/2018-11-13-events.json
+```
+
+And below is an example of what the data in a log file, 2018-11-12-events.json, looks like.
+
+![log-data](assets/log-data.png)
+
+
+
+
+
