--- lessons/cloud-data-warehouses/Data_Warehouse_Project/requirements.txt
@@ -1,5 +1,0 @@
-psycopg2-binary==2.8.6
-boto3==1.16.59
-pandas==1.2.1
-click==7.1.2
-
